Actually fix seeing other players
Needed to store players in the hashmap by UUID instead of name. Dummy
players cannot be seen, only actual players can. This can result in
situations where the only evidence of a player's existence in another
server is through their dummy which the viewer cannot see, but this is
intended behavior.