Filter out invalid mappings in destroy entities
PlayDestroyEntities has a list of entities to destroy. Previously, this
would just not send the packet if there does not exist an eid mapping
for any id in the list, but now we just filter out any eids that don't
have a mapping. Possibly more intensive since we're using a vec, but its
the correct behavior now.